Landing pages redo - Separated Home into Landing1 and Landing2 - Connected About and Home
