+++ src/index.css
@@ -3,7 +3,7 @@
 
 @theme {
   /** Colors **/
-  --color-purple1: #b7bcec; /* light purple: #B7BCEC */
+  --color-purple1: #b7b cec; /* light purple: #B7BCEC */
   --color-purple2: #403b66; /* medium purple #403B66 */
   --color-purple3: #1d1a33; /* dark purple: #1D1A33 */
   --color-purple4: #9e51ce; /* pink purple: #9E51CE */
